+++ apps/vscode-extension/src/preview/Preview.ts
@@ -8,10 +8,11 @@ import {
   setContext,
   togglePresentationView,
 } from '../utils';
-import { DemoRunner, Slides } from '../services';
-import { COMMAND, WebViewMessages, Config } from '@demotime/common';
+import { DemoRunner, DemoStatusBar, Slides } from '../services';
+import { COMMAND, WebViewMessages, Config, Action } from '@demotime/common';
 import { BaseWebview } from '../webview/BaseWebviewPanel';
 import { WebviewType } from '../models';
+import { PresenterView } from '../presenterView/PresenterView';
 
 export class Preview extends BaseWebview {
   public static id: WebviewType = 'preview';
@@ -209,6 +210,7 @@ export class Preview extends BaseWebview {
       setContext(ContextKeys.hasPreviousSlide, payload);
     } else if (command === WebViewMessages.toVscode.nextSlideTitle) {
       Preview.nextSlideTitle = payload;
+      Preview.sendSlideData(payload);
     } else if (command === WebViewMessages.toVscode.openFile && payload) {
       const fileUri = getAbsolutePath(payload);
       await window.showTextDocument(fileUri, { preview: false });
@@ -216,6 +218,33 @@ export class Preview extends BaseWebview {
       Preview.currentSlideIndex = payload;
     } else if (command === WebViewMessages.toVscode.slideReady) {
       Preview.reveal();
+    }
+  }
+
+  private static async sendSlideData(nextSlideTitle?: string) {
+    if (!PresenterView.isOpen) {
+      return;
+    }
+
+    let hasNextSlide = Preview.checkIfHasNextSlide();
+    const demo = hasNextSlide ? DemoRunner.currentDemo : DemoStatusBar.getNextDemo();
+
+    if (!hasNextSlide) {
+      const slideStep = demo?.steps.find((step) => step.action === Action.OpenSlide && step.path);
+      if (slideStep) {
+        hasNextSlide = true;
+      }
+    }
+
+    nextSlideTitle = hasNextSlide ? Preview.getNextSlideTitle() : 'undefined';
+    console.log('Preview sending slide data to PresenterView', { hasNextSlide, nextSlideTitle });
+    if (hasNextSlide) {
+      PresenterView.postMessage(
+        WebViewMessages.toWebview.presenter.nextSlide,
+        nextSlideTitle || '',
+      );
+    } else {
+      PresenterView.postMessage(WebViewMessages.toWebview.presenter.nextSlide, undefined);
     }
   }